requirements.txt
openai==1.40.0
python-telegram-bot==21.4
